refactor: 💡 adjust style and folder structure
