--- apps/e-learning-app/src/app/app.module.css
@@ -1,170 +1,0 @@
-body {
-  overflow-y: overlay;
-  scrollbar-color: #1f4e5a transparent;
-  scrollbar-width: thin;
-}
-
-.app {
-  font-family: sans-serif;
-  min-width: 300px;
-  max-width: 600px;
-  margin: 50px auto;
-}
-
-.app :global(.gutter-left) {
-  margin-left: 9px;
-}
-
-.app :global(.col-span-2) {
-  grid-column: span 2;
-}
-
-.app :global(.flex) {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.app :global(header) {
-  background-color: #143055;
-  color: white;
-  padding: 5px;
-  border-radius: 3px;
-}
-
-.app :global(main) {
-  padding: 0 36px;
-}
-
-.app :global(p) {
-  text-align: center;
-}
-
-.app :global(h1) {
-  text-align: center;
-  margin-left: 18px;
-  font-size: 24px;
-}
-
-.app :global(h2) {
-  text-align: center;
-  font-size: 20px;
-  margin: 40px 0 10px 0;
-}
-
-.app :global(.resources) {
-  text-align: center;
-  list-style: none;
-  padding: 0;
-  display: grid;
-  grid-gap: 9px;
-  grid-template-columns: 1fr 1fr;
-}
-
-.app :global(.resource) {
-  color: #0094ba;
-  height: 36px;
-  background-color: rgba(0, 0, 0, 0);
-  border: 1px solid rgba(0, 0, 0, 0.12);
-  border-radius: 4px;
-  padding: 3px 9px;
-  text-decoration: none;
-}
-
-.app :global(.resource:hover) {
-  background-color: rgba(68, 138, 255, 0.04);
-}
-
-.app :global(pre) {
-  padding: 9px;
-  border-radius: 4px;
-  background-color: black;
-  color: #eee;
-}
-
-.app :global(details) {
-  border-radius: 4px;
-  color: #333;
-  background-color: rgba(0, 0, 0, 0);
-  border: 1px solid rgba(0, 0, 0, 0.12);
-  padding: 3px 9px;
-  margin-bottom: 9px;
-}
-
-.app :global(summary) {
-  outline: none;
-  height: 36px;
-  line-height: 36px;
-}
-
-.app :global(.github-star-container) {
-  margin-top: 12px;
-  line-height: 20px;
-}
-
-.app :global(.github-star-container a) {
-  display: flex;
-  align-items: center;
-  text-decoration: none;
-  color: #333;
-}
-
-.app :global(.github-star-badge) {
-  color: #24292e;
-  display: flex;
-  align-items: center;
-  font-size: 12px;
-  padding: 3px 10px;
-  border: 1px solid rgba(27, 31, 35, 0.2);
-  border-radius: 3px;
-  background-image: linear-gradient(-180deg, #fafbfc, #eff3f6 90%);
-  margin-left: 4px;
-  font-weight: 600;
-}
-
-.app :global(.github-star-badge:hover) {
-  background-image: linear-gradient(-180deg, #f0f3f6, #e6ebf1 90%);
-  border-color: rgba(27, 31, 35, 0.35);
-  background-position: -0.5em;
-}
-.app :global(.github-star-badge .material-icons) {
-  height: 16px;
-  width: 16px;
-  margin-right: 4px;
-}
-
-::-webkit-scrollbar {
-  width: 10px;
-  background-color: transparent;
-}
-
-::-webkit-scrollbar-corner,
-::-webkit-scrollbar-track {
-  background-color: transparent;
-}
-
-::-webkit-scrollbar-thumb {
-  width: 10px;
-  border-radius: 5px;
-  background-color: #1f4e5a;
-  background-clip: padding-box;
-  border: 0.13rem solid transparent;
-}
-
-::-webkit-scrollbar-button:single-button:vertical:decrement {
-  height: 14px;
-  width: 10px;
-  background-position: center;
-  background-image: url('../assets/icons/scroll_arrow_up.svg');
-  background-size: 10.5px 6px;
-  background-repeat: no-repeat;
-}
-
-::-webkit-scrollbar-button:single-button:vertical:increment {
-  height: 14px;
-  width: 10px;
-  background-position: center;
-  background-image: url('../assets/icons/scroll_arrow_down.svg');
-  background-size: 10.5px 6px;
-  background-repeat: no-repeat;
-}
