+++ apps/e-learning-app/src/app/themes/FontFamilies.js
@@ -1,0 +1,8 @@
+const fontFamilies = {
+  body_semibold: {
+    fontFamily: "'Open Sans', sans-serif",
+    fontWeight: 600,
+  }
+}
+
+export default fontFamilies;
