+++ apps/e-learning-app/src/app/themes/Sizes.js
@@ -1,0 +1,2 @@
+export const sidebarButtonHeight = 6.85;
+export const navBarHeight = 6.85;
