+++ apps/e-learning-app/src/app/utilities/MockData.js
