--- apps/e-learning-app/src/app/utilities/UtilityValues.js
@@ -1,5 +1,0 @@
-//Contains constants values used in other files
-
-export const linkSize = 6.85;
-export const navMenuSize = 6.85;
-export const defaultNumber = Math.floor((100 - navMenuSize) / linkSize);
